Move common header to net
