+++ examples/http/client.cc
@@ -1,5 +1,5 @@
 #include "opentelemetry/ext/http/client/http_client_factory.h"
-#include "opentelemetry/ext/http/common/url_parser.h"
+#include "opentelemetry/ext/net/common/url_parser.h"
 #include "tracer_common.h"
 
 namespace
